convert editActions to ts
